refactor: refactoring the code according to yuliaHope requirements
