--- client/src/components/Schedule/CalendarView.tsx
@@ -1,19 +1,0 @@
-import { CourseEvent } from 'services/course';
-
-type Props = {
-  data: CourseEvent[];
-  timeZone: string;
-};
-
-
-export function CalendarView({ data, timeZone }: Props) {
-  console.log(data);
-  return (
-    <div>
-      <h2>Calendar</h2>
-      Data length = {data.length}. Timezone = {timeZone}
-    </div>
-  );
-}
-
-export default CalendarView;
