+++ client/src/configs/timezones.ts
@@ -5,7 +5,7 @@ export const TIMEZONES = [
   'Europe/Minsk',
   'Europe/Moscow',
   'Europe/Volgograd',
-  'Europe/Yekaterinburg',
+  'Asia/Yekaterinburg',
   'Asia/Tashkent',
   'Asia/Tbilisi',
 ];
