--- server/.env.example
@@ -1,22 +1,0 @@
-TZ=utc
-
-RSSHCOOL_API_PG_HOST=localhost
-RSSHCOOL_API_PG_USERNAME=rs_master
-RSSHCOOL_API_PG_PASSWORD=12345678
-RSSHCOOL_API_PG_DATABASE=rs_school
-
-RSSHCOOL_API_ADMIN_USERNAME=admin
-RSSHCOOL_API_ADMIN_PASSWORD=password
-
-# set to false to disable "admin" mode
-RSSHCOOL_DEV_ADMIN=true
-
-# mentors 2020 q1
-# RSSHCOOL_DEV_USERNAME=mikhama
-# RSSHCOOL_DEV_USERNAME=dmitryromaniuk
-
-# students 2020 q1
-# RSSHCOOL_DEV_USERNAME=apalchys
-RSSHCOOL_DEV_USERNAME=yuliahope
-
-
